changed migrations id column type from incremets to bigIncremets and change integer type to bigInteger; applied those changes for test cases;

+++ tests/UsedByTeamsTraitTest.php
@@ -20,16 +20,16 @@ class UsedByTeamsTraitTest extends Orchestra\Testbench\TestCase
         $app['config']->set('teamwork.user_model', 'User');
 
         \Schema::create('users', function ($table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->timestamps();
             $table->softDeletes();
         });
 
         \Schema::create('tasks', function ($table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('team_id');
+            $table->bigInteger('team_id');
             $table->timestamps();
         });
     }
